+++ photutils/utils/__init__.py
@@ -2,8 +2,7 @@
 """
 General-purpose utility functions.
 """
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
 from .check_random_state import *
+from .colormaps import *
 from .interpolation import *
 from .prepare_data import *